requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.3
bio==1.7.1
biopython==1.85
biothings_client==0.4.1
black==25.1.0
boto3==1.37.18
botocore==1.37.18
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chess==1.11.2
click==8.1.8
cobble==0.1.4
cryptography==44.0.2
dataclasses-json==0.6.7
datasets==3.4.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
duckduckgo_search==7.5.3
et_xmlfile==2.0.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
frozenlist==1.5.0
fsspec==2024.12.0
google_search_results==2.4.2
gprofiler-official==1.0.0
gradio==5.22.0
gradio_client==1.8.0
greenlet==3.1.1
groovy==0.1.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.47
langchain-text-splitters==0.3.7
langsmith==0.3.18
litellm==1.63.14
lxml==5.3.1
mammoth==1.9.0
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.16
mygene==3.2.2
mypy-extensions==1.0.0
networkx==3.4.2
numexpr==2.10.2
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.68.2
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pathvalidate==3.2.3
pdfminer==20191125
pdfminer.six==20240706
pillow==11.1.0
platformdirs==4.3.7
pooch==1.8.2
primp==0.14.0
propcache==0.3.0
PubChemPy==1.0.4
puremagic==1.28
pyarrow==19.0.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pypdf==5.4.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
ruff==0.11.2
s3transfer==0.11.4
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
serpapi==0.1.5
setuptools==77.0.3
shellingham==1.5.4
six==1.17.0
smolagents==1.12.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.14.2
SQLAlchemy==2.0.39
starlette==0.46.1
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.50.0
triton==3.2.0
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0.1
wikipedia==1.4.0
xlrd==2.0.1
XlsxWriter==3.2.2
xxhash==3.5.0
yarl==1.18.3
youtube-transcript-api==1.0.2
zipp==3.21.0
zstandard==0.23.0
